requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6